fix(dns): allow UTF-8 TXT content, return integer last_page, read the real SOA ttl key

+++ lib/Application/Controller/Api/V1/UsersController.php
@@ -369,7 +369,7 @@ class UsersController extends PublicApiController
                         'current_page' => $page,
                         'per_page' => $perPage,
                         'total' => $totalCount,
-                        'last_page' => ceil($totalCount / $perPage)
+                        'last_page' => (int)ceil($totalCount / $perPage)
                     ],
                     'meta' => [
                         'timestamp' => date('Y-m-d H:i:s')
